requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
pg8000==1.31.2
pytest==8.3.2
gunicorn==22.0.0
